E-14: extract InterfaceImplEmitter from ReflectionMetadataEmitter (#1361)
Moves the interface method-implementation band (the trailing ~590 lines
of ReflectionMetadataEmitter.cs) into a new internal sealed
InterfaceImplEmitter collaborator per ADR-0150:

- EmitExplicitInterfaceMethodImpls (issues #985/#2010/#2181)
- EmitExplicitInterfacePropertyMethodImpls (issue #2362)
- EmitExplicitInterfaceEventMethodImpls (ADR-0149)
- EmitStaticVirtualMethodImpls (ADR-0089, issues #755/#1268/#2370)
- EmitStaticVirtualPropertyMethodImpls (ADR-0089, issue #1019)
- Private helpers used only by this band: ResolveOpenInterfaceInstanceMethod,
  ResolveOpenInterfaceStaticMethod, StaticVirtualSignatureEquals

Method bodies moved verbatim; the only rewrites are mechanical reference
retargets (this.GetMethodReference -> this.outer.GetMethodReference,
this.ResolveUserInterfaceInstanceMethodToken ->
this.outer.ResolveUserInterfaceInstanceMethodToken,
IsUserGenericInterfaceReference ->
ReflectionMetadataEmitter.IsUserGenericInterfaceReference) and visibility
lifts on the moved members themselves (band entry points private ->
internal; ResolveOpenInterfaceStaticMethod private static -> internal
static for the root's remaining caller in
ResolveUserInterfaceStaticMethodToken).

Wiring: back-reference ctor `InterfaceImplEmitter(ReflectionMetadataEmitter
outer)` (the MethodBodyEmitter idiom), reading emitCtx/cache off the
back-reference into direct fields; constructed in EmitCore right after
assemblyAttrs (the E-13 peer).

Widenings: none — every root member the band depends on
(GetMethodReference, ResolveUserInterfaceInstanceMethodToken,
IsUserGenericInterfaceReference, emitCtx, cache) was already internal
from PR-E-11/E-12.

Forwarders: none needed — grep across the Emit folder (MethodBodyEmitter
partials, TypeDefEmitter, StateMachineEmitter, ClosureEmitter,
MemberDefEmitter) found no callers of the moved methods outside RME's
own remaining code; the 11 RME call sites were repointed directly.
(MethodBodyEmitter.Calls.cs's ResolveOpenInterfaceMethod is a distinct
sibling copy and is untouched.)

Verification: dotnet build GSharp.sln -c Release -graph = 0 warnings /
0 errors; samples corpus PE byte-diff 119/119 identical (3 pre-existing
baseline failures: GsharpExtensionsMixed/Optional/Sequences need
/r: Gsharp.Extensions).
