+++ src/Core/CodeAnalysis/Emit/ReflectionMetadataEmitter.cs
@@ -116,6 +116,16 @@ internal sealed class ReflectionMetadataEmitter
     // EmitCore-ordering reason as the other E-* components.
     private AssemblyAttributeEmitter assemblyAttrs;
 
+    // PR-E-14: InterfaceImplEmitter — owns the MethodImpl-row band that used
+    // to close out this file (EmitExplicitInterfaceMethodImpls /
+    // EmitExplicitInterfacePropertyMethodImpls /
+    // EmitExplicitInterfaceEventMethodImpls / EmitStaticVirtualMethodImpls /
+    // EmitStaticVirtualPropertyMethodImpls) plus the open-slot resolvers and
+    // static-virtual signature matcher used only by that band. Back-bound to
+    // this root (the MethodBodyEmitter idiom) for GetMethodReference /
+    // ResolveUserInterfaceInstanceMethodToken / IsUserGenericInterfaceReference.
+    private InterfaceImplEmitter interfaceImpls;
+
     // Phase 4 emit parity (E1): synthesized lambda bodies (no captures).
     // Populated by a pre-pass walker over every user function/entry body.
     // Each lambda's synthetic FunctionSymbol is registered alongside user
@@ -934,6 +944,14 @@ internal sealed class ReflectionMetadataEmitter
             this.wellKnown,
             this.customAttrEncoder,
             this.GetTypeReference);
+
+        // PR-E-14: InterfaceImplEmitter owns the explicit-interface and
+        // static-virtual MethodImpl-row emitters. Back-bound to this root
+        // (the MethodBodyEmitter idiom) because its token resolution runs
+        // through GetMethodReference / ResolveUserInterfaceInstanceMethodToken,
+        // which stay on the root; it reads emitCtx/cache directly off the
+        // back-reference in its ctor.
+        this.interfaceImpls = new InterfaceImplEmitter(this);
 
         // PR-E-12: MethodBodyPlanner owns the per-body planning orchestrators
         // (CollectLocalsAndLabels and friends) that drive SlotPlanner's
@@ -3097,23 +3115,23 @@ internal sealed class ReflectionMetadataEmitter
 
             // ADR-0089 / issue #755: emit MethodImpl rows for static-virtual
             // interface members. See structs path for the same call.
-            this.EmitStaticVirtualMethodImpls(c);
+            this.interfaceImpls.EmitStaticVirtualMethodImpls(c);
 
             // ADR-0089 / issue #1019: emit MethodImpl rows for static-virtual
             // interface properties (accessor methods).
-            this.EmitStaticVirtualPropertyMethodImpls(c);
+            this.interfaceImpls.EmitStaticVirtualPropertyMethodImpls(c);
 
             // Issue #985: emit MethodImpl rows for covariant-return interface
             // bridges (e.g. the non-generic IEnumerable.GetEnumerator).
-            this.EmitExplicitInterfaceMethodImpls(c);
+            this.interfaceImpls.EmitExplicitInterfaceMethodImpls(c);
 
             // Issue #2362: emit MethodImpl rows for mangled-name explicit
             // interface property implementations (accessor methods).
-            this.EmitExplicitInterfacePropertyMethodImpls(c);
+            this.interfaceImpls.EmitExplicitInterfacePropertyMethodImpls(c);
 
             // ADR-0149: emit MethodImpl rows for explicit-interface-clause
             // event implementations (add/remove/raise accessors).
-            this.EmitExplicitInterfaceEventMethodImpls(c);
+            this.interfaceImpls.EmitExplicitInterfaceEventMethodImpls(c);
         }
 
         foreach (var c in topClasses)
@@ -3199,23 +3217,23 @@ internal sealed class ReflectionMetadataEmitter
             // required so `constrained.` dispatch finds the right body. The
             // row points the interface slot's MethodDef at the implementer's
             // static MethodDef on the struct's TypeDef.
-            this.EmitStaticVirtualMethodImpls(s);
+            this.interfaceImpls.EmitStaticVirtualMethodImpls(s);
 
             // ADR-0089 / issue #1019: emit MethodImpl rows for static-virtual
             // interface properties (accessor methods).
-            this.EmitStaticVirtualPropertyMethodImpls(s);
+            this.interfaceImpls.EmitStaticVirtualPropertyMethodImpls(s);
 
             // Issue #985: emit MethodImpl rows for covariant-return interface
             // bridges declared on a struct that implements `IEnumerable[T]` &c.
-            this.EmitExplicitInterfaceMethodImpls(s);
+            this.interfaceImpls.EmitExplicitInterfaceMethodImpls(s);
 
             // Issue #2362: emit MethodImpl rows for mangled-name explicit
             // interface property implementations (accessor methods).
-            this.EmitExplicitInterfacePropertyMethodImpls(s);
+            this.interfaceImpls.EmitExplicitInterfacePropertyMethodImpls(s);
 
             // ADR-0149: emit MethodImpl rows for explicit-interface-clause
             // event implementations (add/remove/raise accessors).
-            this.EmitExplicitInterfaceEventMethodImpls(s);
+            this.interfaceImpls.EmitExplicitInterfaceEventMethodImpls(s);
         }
 
         foreach (var s in topStructs)
@@ -8186,7 +8204,7 @@ internal sealed class ReflectionMetadataEmitter
     /// <returns>The MethodDef or TypeSpec-parented MemberRef token.</returns>
     internal EntityHandle ResolveUserInterfaceStaticMethodToken(InterfaceSymbol containingInterface, FunctionSymbol method)
     {
-        var openMethod = ResolveOpenInterfaceStaticMethod(containingInterface, method);
+        var openMethod = InterfaceImplEmitter.ResolveOpenInterfaceStaticMethod(containingInterface, method);
         if (!this.cache.MethodHandles.TryGetValue(openMethod, out var openDef)
             && !this.cache.FunctionHandles.TryGetValue(openMethod, out openDef))
         {
@@ -12010,592 +12028,4 @@ internal sealed class ReflectionMetadataEmitter
     // PR-E-11: BodyEmitter promoted to top-level MethodBodyEmitter
     // (src/Core/CodeAnalysis/Emit/MethodBodyEmitter.cs and partials).
     // PR-E-2: MethodSpecSymbolKey moved into MetadataTokenCache.
-
-    /// <summary>
-    /// Issue #985 / #2010: emit MethodImpl rows for covariant-return interface
-    /// bridges AND for mangled-name explicit interface implementations. A
-    /// method whose <see cref="FunctionSymbol.ExplicitInterfaceSlot"/> is set
-    /// explicitly implements a specific (typically inherited, non-generic) CLR
-    /// interface slot — e.g. the private non-generic
-    /// <c>IEnumerable.GetEnumerator()</c> alongside the public generic
-    /// <c>IEnumerable&lt;T&gt;.GetEnumerator()</c>. A private bridge method
-    /// cannot implicitly implement an interface slot, so the explicit row is
-    /// required for the resulting type to load.
-    ///
-    /// A method whose <see cref="FunctionSymbol.ExplicitInterfaceMember"/> is
-    /// set explicitly implements one specific in-compilation (G#) interface
-    /// member — its mangled name never matches the interface member's own
-    /// name, so ordinary name-based virtual dispatch never wires it into that
-    /// interface's slot; an explicit <c>MethodImpl</c> row is required here
-    /// too (mirrors <see cref="EmitStaticVirtualMethodImpls"/>'s generic-aware
-    /// token resolution for a constructed interface).
-    /// </summary>
-    /// <param name="structSymbol">The implementing class or struct.</param>
-    private void EmitExplicitInterfaceMethodImpls(StructSymbol structSymbol)
-    {
-        if (structSymbol == null || structSymbol.Methods.IsDefaultOrEmpty)
-        {
-            return;
-        }
-
-        if (!this.cache.StructTypeDefs.TryGetValue(structSymbol, out var implTypeDef))
-        {
-            return;
-        }
-
-        foreach (var method in structSymbol.Methods)
-        {
-            if (!this.cache.MethodHandles.TryGetValue(method, out var implHandle))
-            {
-                continue;
-            }
-
-            var slot = method.ExplicitInterfaceSlot;
-            if (slot != null)
-            {
-                var slotRef = this.GetMethodReference(slot);
-                this.emitCtx.Metadata.AddMethodImplementation(implTypeDef, implHandle, slotRef);
-            }
-
-            var ifaceMember = method.ExplicitInterfaceMember;
-            if (ifaceMember == null || structSymbol.Interfaces.IsDefaultOrEmpty)
-            {
-                continue;
-            }
-
-            EntityHandle? slotHandle = null;
-            foreach (var iface in structSymbol.Interfaces)
-            {
-                // Issue #2181: for a constructed generic interface, the linked
-                // `ExplicitInterfaceMember` is the SUBSTITUTED method surfaced
-                // on the constructed instance (`iface.Methods`), not the open
-                // definition's slot — so match against the constructed
-                // instance's methods here, then map back to the open slot for
-                // the MethodDef / MemberRef token resolution (the open slot is
-                // what `MethodHandles` and `ResolveUserInterfaceInstanceMethodToken`
-                // are keyed by). For a non-generic interface `iface.Methods` is
-                // the definition's own table and the map is an identity, so the
-                // pre-existing behavior is unchanged.
-                if (iface.Methods.IsDefaultOrEmpty || !iface.Methods.Contains(ifaceMember))
-                {
-                    continue;
-                }
-
-                var openSlot = ResolveOpenInterfaceInstanceMethod(iface, ifaceMember);
-                slotHandle = IsUserGenericInterfaceReference(iface)
-                    ? this.ResolveUserInterfaceInstanceMethodToken(iface, openSlot)
-                    : this.cache.MethodHandles.TryGetValue(openSlot, out var slotDefHandle)
-                        ? slotDefHandle
-                        : (EntityHandle?)null;
-                break;
-            }
-
-            if (slotHandle.HasValue)
-            {
-                this.emitCtx.Metadata.AddMethodImplementation(implTypeDef, implHandle, slotHandle.Value);
-            }
-        }
-    }
-
-    /// <summary>
-    /// Issue #2362: emit <c>MethodImpl</c> rows for mangled-name explicit
-    /// interface PROPERTY implementations — the property-level counterpart of
-    /// <see cref="EmitExplicitInterfaceMethodImpls"/>. A property whose
-    /// <see cref="PropertySymbol.ExplicitInterfaceMember"/> is set explicitly
-    /// implements one specific in-compilation (G#) interface property; its
-    /// mangled name never matches the interface member's own name, so
-    /// ordinary name-based virtual dispatch never wires its accessors into
-    /// that interface's slot. An explicit <c>MethodImpl</c> row per accessor
-    /// (getter and/or setter) is required, exactly mirroring
-    /// <see cref="EmitStaticVirtualPropertyMethodImpls"/>'s generic-aware
-    /// token resolution for a constructed interface, but for an INSTANCE
-    /// property slot resolved from <see cref="StructSymbol.Interfaces"/>
-    /// rather than a static-virtual slot.
-    /// </summary>
-    /// <param name="structSymbol">The implementing class or struct.</param>
-    private void EmitExplicitInterfacePropertyMethodImpls(StructSymbol structSymbol)
-    {
-        if (structSymbol == null || structSymbol.Properties.IsDefaultOrEmpty || structSymbol.Interfaces.IsDefaultOrEmpty)
-        {
-            return;
-        }
-
-        if (!this.cache.StructTypeDefs.TryGetValue(structSymbol, out var implTypeDef))
-        {
-            return;
-        }
-
-        foreach (var prop in structSymbol.Properties)
-        {
-            var ifaceMember = prop.ExplicitInterfaceMember;
-            if (ifaceMember == null)
-            {
-                continue;
-            }
-
-            if (!this.cache.PropertyAccessorHandles.TryGetValue(prop, out var implAccessors))
-            {
-                continue;
-            }
-
-            foreach (var iface in structSymbol.Interfaces)
-            {
-                // Issue #2362: InterfaceSymbol.Construct does not substitute
-                // Properties onto a constructed generic instance (see
-                // InterfaceSymbol.TryResolveMembers) — `ExplicitInterfaceMember`
-                // is always linked against the OPEN definition's property (see
-                // TryResolveExplicitInterfacePropertyImplementation), so match
-                // against `defIface.Properties` here too, exactly like
-                // EmitStaticVirtualPropertyMethodImpls.
-                var defIface = iface.Definition ?? iface;
-                if (defIface.Properties.IsDefaultOrEmpty || !defIface.Properties.Contains(ifaceMember))
-                {
-                    continue;
-                }
-
-                var isGenericIface = IsUserGenericInterfaceReference(iface);
-
-                if (ifaceMember.HasGetter && implAccessors.Getter.HasValue && ifaceMember.GetterSymbol != null)
-                {
-                    EntityHandle? getterDecl = isGenericIface
-                        ? this.ResolveUserInterfaceInstanceMethodToken(iface, ifaceMember.GetterSymbol)
-                        : this.cache.MethodHandles.TryGetValue(ifaceMember.GetterSymbol, out var getterDefHandle)
-                            ? getterDefHandle
-                            : (EntityHandle?)null;
-                    if (getterDecl.HasValue)
-                    {
-                        this.emitCtx.Metadata.AddMethodImplementation(implTypeDef, implAccessors.Getter.Value, getterDecl.Value);
-                    }
-                }
-
-                if (ifaceMember.HasSetter && implAccessors.Setter.HasValue && ifaceMember.SetterSymbol != null)
-                {
-                    EntityHandle? setterDecl = isGenericIface
-                        ? this.ResolveUserInterfaceInstanceMethodToken(iface, ifaceMember.SetterSymbol)
-                        : this.cache.MethodHandles.TryGetValue(ifaceMember.SetterSymbol, out var setterDefHandle)
-                            ? setterDefHandle
-                            : (EntityHandle?)null;
-                    if (setterDecl.HasValue)
-                    {
-                        this.emitCtx.Metadata.AddMethodImplementation(implTypeDef, implAccessors.Setter.Value, setterDecl.Value);
-                    }
-                }
-
-                break;
-            }
-        }
-    }
-
-    /// <summary>
-    /// ADR-0149: emit <c>MethodImpl</c> rows for explicit-interface-clause
-    /// EVENT implementations (<c>event (IFoo) Changed T</c>) — the event-level
-    /// counterpart of <see cref="EmitExplicitInterfacePropertyMethodImpls"/>,
-    /// generalizing the #2362 explicit-implementation convention to events
-    /// for the first time. An event whose
-    /// <see cref="EventSymbol.ExplicitInterfaceMember"/> is set explicitly
-    /// implements one specific in-compilation interface event; its add/
-    /// remove (and, if present, raise) accessors are bridged into that
-    /// interface's abstract slots via one <c>MethodImpl</c> row per accessor,
-    /// exactly mirroring the property function's generic-aware token
-    /// resolution.
-    /// </summary>
-    /// <param name="structSymbol">The implementing class or struct.</param>
-    private void EmitExplicitInterfaceEventMethodImpls(StructSymbol structSymbol)
-    {
-        if (structSymbol == null || structSymbol.Events.IsDefaultOrEmpty || structSymbol.Interfaces.IsDefaultOrEmpty)
-        {
-            return;
-        }
-
-        if (!this.cache.StructTypeDefs.TryGetValue(structSymbol, out var implTypeDef))
-        {
-            return;
-        }
-
-        foreach (var ev in structSymbol.Events)
-        {
-            var ifaceMember = ev.ExplicitInterfaceMember;
-            if (ifaceMember == null)
-            {
-                continue;
-            }
-
-            if (!this.cache.EventAccessorHandles.TryGetValue(ev, out var implAccessors))
-            {
-                continue;
-            }
-
-            foreach (var iface in structSymbol.Interfaces)
-            {
-                // ADR-0149 (mirrors #2362's property resolution): interface
-                // Events, like Properties, are never substituted onto a
-                // constructed generic instance (InterfaceSymbol.TryResolveMembers),
-                // so ExplicitInterfaceMember is always linked against the OPEN
-                // definition's event — match against `defIface.Events` here.
-                var defIface = iface.Definition ?? iface;
-                if (defIface.Events.IsDefaultOrEmpty || !defIface.Events.Contains(ifaceMember))
-                {
-                    continue;
-                }
-
-                var isGenericIface = IsUserGenericInterfaceReference(iface);
-
-                if (ifaceMember.AddMethodSymbol != null)
-                {
-                    EntityHandle? addDecl = isGenericIface
-                        ? this.ResolveUserInterfaceInstanceMethodToken(iface, ifaceMember.AddMethodSymbol)
-                        : this.cache.MethodHandles.TryGetValue(ifaceMember.AddMethodSymbol, out var addDefHandle)
-                            ? addDefHandle
-                            : (EntityHandle?)null;
-                    if (addDecl.HasValue)
-                    {
-                        this.emitCtx.Metadata.AddMethodImplementation(implTypeDef, implAccessors.Add, addDecl.Value);
-                    }
-                }
-
-                if (ifaceMember.RemoveMethodSymbol != null)
-                {
-                    EntityHandle? removeDecl = isGenericIface
-                        ? this.ResolveUserInterfaceInstanceMethodToken(iface, ifaceMember.RemoveMethodSymbol)
-                        : this.cache.MethodHandles.TryGetValue(ifaceMember.RemoveMethodSymbol, out var removeDefHandle)
-                            ? removeDefHandle
-                            : (EntityHandle?)null;
-                    if (removeDecl.HasValue)
-                    {
-                        this.emitCtx.Metadata.AddMethodImplementation(implTypeDef, implAccessors.Remove, removeDecl.Value);
-                    }
-                }
-
-                if (ifaceMember.RaiseMethodSymbol != null && implAccessors.Raise.HasValue)
-                {
-                    EntityHandle? raiseDecl = isGenericIface
-                        ? this.ResolveUserInterfaceInstanceMethodToken(iface, ifaceMember.RaiseMethodSymbol)
-                        : this.cache.MethodHandles.TryGetValue(ifaceMember.RaiseMethodSymbol, out var raiseDefHandle)
-                            ? raiseDefHandle
-                            : (EntityHandle?)null;
-                    if (raiseDecl.HasValue)
-                    {
-                        this.emitCtx.Metadata.AddMethodImplementation(implTypeDef, implAccessors.Raise.Value, raiseDecl.Value);
-                    }
-                }
-
-                break;
-            }
-        }
-    }
-
-    /// <summary>
-    /// ADR-0089 / issue #755: emit MethodImpl rows binding each declared
-    /// static-virtual interface slot to the implementer's matching static
-    /// method on <paramref name="structSymbol"/>. Best-effort match — if
-    /// the implementer is missing the slot, the binder has already issued
-    /// GS0331/GS0332 and we skip silently here.
-    /// </summary>
-    private void EmitStaticVirtualMethodImpls(StructSymbol structSymbol)
-    {
-        if (structSymbol == null || structSymbol.Interfaces.IsDefaultOrEmpty || structSymbol.StaticMethods.IsDefaultOrEmpty)
-        {
-            return;
-        }
-
-        if (!this.cache.StructTypeDefs.TryGetValue(structSymbol, out var implTypeDef))
-        {
-            return;
-        }
-
-        foreach (var iface in structSymbol.Interfaces)
-        {
-            if (iface.StaticMethods.IsDefaultOrEmpty)
-            {
-                continue;
-            }
-
-            // Issue #1268: when the implemented interface is a constructed
-            // generic (e.g. `TrackNumber : IData[TrackNumber]`), its
-            // `StaticMethods` are substituted instances that are NOT keyed in
-            // `MethodHandles`; only the open definition's slots are. Resolve
-            // each substituted slot back to its open counterpart for the
-            // MethodDef lookup, and parent the MethodImpl's declaration at the
-            // constructed interface's TypeSpec (a MemberRef) so the runtime can
-            // pair the override against `IData`1<TrackNumber>::Method`.
-            var isGenericIface = IsUserGenericInterfaceReference(iface);
-            foreach (var slot in iface.StaticMethods)
-            {
-                var openSlot = ResolveOpenInterfaceStaticMethod(iface, slot);
-                if (!this.cache.MethodHandles.TryGetValue(openSlot, out var slotDefHandle))
-                {
-                    continue;
-                }
-
-                EntityHandle slotHandle = isGenericIface
-                    ? this.ResolveUserInterfaceInstanceMethodToken(iface, openSlot)
-                    : slotDefHandle;
-
-                // ADR-0149 follow-up (issue #2370): an explicit-interface-
-                // clause static method (`func (IFoo) M(...)` inside a
-                // `shared { }` block) is linked by the binder via
-                // `ExplicitInterfaceMember` (DeclarationBinder's
-                // `TryResolveExplicitInterfaceStaticImplementation`) against
-                // this exact constructed-instance `slot`, regardless of its
-                // own declared name — prefer that link over the name-based
-                // `StaticVirtualSignatureEquals` scan below, mirroring
-                // `EmitExplicitInterfaceMethodImpls`'s instance-method
-                // precedent.
-                FunctionSymbol implMatch = null;
-                if (!structSymbol.StaticMethods.IsDefaultOrEmpty)
-                {
-                    foreach (var explicitCandidate in structSymbol.StaticMethods)
-                    {
-                        if (ReferenceEquals(explicitCandidate.ExplicitInterfaceMember, slot))
-                        {
-                            implMatch = explicitCandidate;
-                            break;
-                        }
-                    }
-                }
-
-                if (implMatch == null)
-                {
-                    foreach (var candidate in structSymbol.GetStaticMethods(slot.Name))
-                    {
-                        if (StaticVirtualSignatureEquals(slot, candidate))
-                        {
-                            implMatch = candidate;
-                            break;
-                        }
-                    }
-                }
-
-                if (implMatch == null)
-                {
-                    continue;
-                }
-
-                if (!this.cache.MethodHandles.TryGetValue(implMatch, out var implHandle))
-                {
-                    continue;
-                }
-
-                this.emitCtx.Metadata.AddMethodImplementation(implTypeDef, implHandle, slotHandle);
-            }
-        }
-    }
-
-    /// <summary>
-    /// Issue #1268: maps a (possibly substituted) static-virtual method slot
-    /// on a constructed generic interface back to its open declaration on the
-    /// interface definition. Returns <paramref name="slot"/> unchanged when the
-    /// interface is not a constructed instance, or when no open counterpart is
-    /// found.
-    /// </summary>
-    /// <summary>
-    /// Issue #2181: maps a (possibly substituted) instance method slot on a
-    /// constructed generic interface back to its open declaration on the
-    /// interface definition — the counterpart of
-    /// <see cref="ResolveOpenInterfaceStaticMethod"/> for the (non-static)
-    /// <see cref="InterfaceSymbol.Methods"/> table. Positional match first
-    /// (the constructed instance's <c>Methods</c> mirror the definition's
-    /// order), with a name/arity fallback; returns <paramref name="slot"/>
-    /// unchanged when the interface is a plain definition (identity map).
-    /// </summary>
-    private static FunctionSymbol ResolveOpenInterfaceInstanceMethod(InterfaceSymbol iface, FunctionSymbol slot)
-    {
-        var def = iface.Definition ?? iface;
-        // Function/interface symbols are canonical within one emit pass; CLR
-        // metadata Type identity is not involved in this positional map.
-        if (ReferenceEquals(def, iface))
-        {
-            return slot;
-        }
-
-        var constructedMethods = iface.Methods;
-        var defMethods = def.Methods;
-        for (var i = 0; i < constructedMethods.Length; i++)
-        {
-            if (ReferenceEquals(constructedMethods[i], slot) && i < defMethods.Length)
-            {
-                return defMethods[i];
-            }
-        }
-
-        foreach (var m in defMethods)
-        {
-            if (m.Name == slot.Name && m.Parameters.Length == slot.Parameters.Length)
-            {
-                return m;
-            }
-        }
-
-        return slot;
-    }
-
-    private static FunctionSymbol ResolveOpenInterfaceStaticMethod(InterfaceSymbol iface, FunctionSymbol slot)
-    {
-        var def = iface.Definition ?? iface;
-        // Function/interface symbols are canonical within one emit pass; CLR
-        // metadata Type identity is not involved in this positional map.
-        if (ReferenceEquals(def, iface))
-        {
-            return slot;
-        }
-
-        var constructedStatics = iface.StaticMethods;
-        var defStatics = def.StaticMethods;
-        for (var i = 0; i < constructedStatics.Length; i++)
-        {
-            if (ReferenceEquals(constructedStatics[i], slot) && i < defStatics.Length)
-            {
-                return defStatics[i];
-            }
-        }
-
-        foreach (var m in defStatics)
-        {
-            if (m.Name == slot.Name && m.Parameters.Length == slot.Parameters.Length)
-            {
-                return m;
-            }
-        }
-
-        return slot;
-    }
-
-    /// <summary>
-    /// ADR-0089 / issue #1019: emit <c>MethodImpl</c> rows pairing the
-    /// implementer's static property accessor methods (<c>get_Name</c> /
-    /// <c>set_Name</c>) to the matching static-virtual interface property
-    /// accessor slots. Mirrors <see cref="EmitStaticVirtualMethodImpls"/> but
-    /// resolves accessor MethodDef handles through
-    /// <c>PropertyAccessorHandles</c> (static properties are tracked on
-    /// <see cref="StructSymbol.StaticProperties"/>, not <c>StaticMethods</c>).
-    /// </summary>
-    /// <param name="structSymbol">The implementer struct/class.</param>
-    private void EmitStaticVirtualPropertyMethodImpls(StructSymbol structSymbol)
-    {
-        if (structSymbol == null || structSymbol.Interfaces.IsDefaultOrEmpty || structSymbol.StaticProperties.IsDefaultOrEmpty)
-        {
-            return;
-        }
-
-        if (!this.cache.StructTypeDefs.TryGetValue(structSymbol, out var implTypeDef))
-        {
-            return;
-        }
-
-        foreach (var iface in structSymbol.Interfaces)
-        {
-            // Issue #1268: a constructed generic interface does not surface its
-            // declared properties on the constructed instance (only methods are
-            // substituted) — walk the open definition's property table so the
-            // static-virtual property slots are found, and parent the
-            // MethodImpl declaration at the constructed TypeSpec for generic
-            // interfaces.
-            var defIface = iface.Definition ?? iface;
-            if (defIface.Properties.IsDefaultOrEmpty)
-            {
-                continue;
-            }
-
-            var isGenericIface = IsUserGenericInterfaceReference(iface);
-            foreach (var slotProp in defIface.Properties)
-            {
-                if (!slotProp.IsStatic)
-                {
-                    continue;
-                }
-
-                if (!this.cache.PropertyAccessorHandles.TryGetValue(slotProp, out var slotAccessors))
-                {
-                    continue;
-                }
-
-                // ADR-0149 follow-up (issue #2370): an explicit-interface-
-                // clause static property (`prop (IFoo) P T` inside a
-                // `shared { }` block) is linked by the binder via
-                // `ExplicitInterfaceMember` against this exact `slotProp`
-                // (both resolved from the OPEN interface definition's
-                // `Properties` table — see `VerifyStaticVirtualInterfaceProperty
-                // Implementations`'s `staticPropDefIface` fix) — prefer that
-                // link over the name-based scan below.
-                PropertySymbol implProp = null;
-                foreach (var explicitCandidate in structSymbol.StaticProperties)
-                {
-                    if (ReferenceEquals(explicitCandidate.ExplicitInterfaceMember, slotProp))
-                    {
-                        implProp = explicitCandidate;
-                        break;
-                    }
-                }
-
-                if (implProp == null)
-                {
-                    foreach (var candidate in structSymbol.StaticProperties)
-                    {
-                        // PropertySymbol.Type is a compiler TypeSymbol, not a CLR
-                        // reflection Type; keep symbol identity plus name fallback.
-                        if (candidate.Name == slotProp.Name
-                            && (ReferenceEquals(candidate.Type, slotProp.Type) || candidate.Type?.Name == slotProp.Type?.Name))
-                        {
-                            implProp = candidate;
-                            break;
-                        }
-                    }
-                }
-
-                if (implProp == null
-                    || !this.cache.PropertyAccessorHandles.TryGetValue(implProp, out var implAccessors))
-                {
-                    continue;
-                }
-
-                if (slotProp.HasGetter && slotAccessors.Getter.HasValue && implAccessors.Getter.HasValue)
-                {
-                    EntityHandle getterDecl = isGenericIface && slotProp.GetterSymbol != null
-                        ? this.ResolveUserInterfaceInstanceMethodToken(iface, slotProp.GetterSymbol)
-                        : slotAccessors.Getter.Value;
-                    this.emitCtx.Metadata.AddMethodImplementation(implTypeDef, implAccessors.Getter.Value, getterDecl);
-                }
-
-                if (slotProp.HasSetter && slotAccessors.Setter.HasValue && implAccessors.Setter.HasValue)
-                {
-                    EntityHandle setterDecl = isGenericIface && slotProp.SetterSymbol != null
-                        ? this.ResolveUserInterfaceInstanceMethodToken(iface, slotProp.SetterSymbol)
-                        : slotAccessors.Setter.Value;
-                    this.emitCtx.Metadata.AddMethodImplementation(implTypeDef, implAccessors.Setter.Value, setterDecl);
-                }
-            }
-        }
-    }
-
-    private static bool StaticVirtualSignatureEquals(FunctionSymbol a, FunctionSymbol b)
-    {
-        if (a == null || b == null)
-        {
-            return false;
-        }
-
-        if (a.Parameters.Length != b.Parameters.Length)
-        {
-            return false;
-        }
-
-        // FunctionSymbol.Type and parameter Type values are compiler symbols
-        // canonicalized for this emit pass, not reflection Type instances.
-        if (!ReferenceEquals(a.Type, b.Type) && a.Type?.Name != b.Type?.Name)
-        {
-            return false;
-        }
-
-        for (var i = 0; i < a.Parameters.Length; i++)
-        {
-            var pa = a.Parameters[i].Type;
-            var pb = b.Parameters[i].Type;
-            if (!ReferenceEquals(pa, pb) && pa?.Name != pb?.Name)
-            {
-                return false;
-            }
-        }
-
-        return true;
-    }
 }
